refactor: RankPrinter를 view 패키지로 변경
