+++ src/test/kotlin/view/RankPrinterTest.kt
@@ -1,5 +1,7 @@
-package lottery.domain
+package view
 
+import lottery.domain.Rank
+import lottery.view.RankPrinter
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
